@@ -1,31 +1,18 @@
 import * as React from "react"
 import { Input, Row, Col } from "antd"
 
-export default class InputValue extends React.Component {
-  state = {
-    value: ""
-  }
+export function InputValue(props) {
+  const [value, setValue] = React.useState("")
+  const onChange = React.useCallback(e => setValue(e.target.value), [])
 
-  onChange = e => this.setState({ value: e.target.value })
-
-  render() {
-    const { value } = this.state
-
-    return (
-      <React.Fragment>
-        <Row>
-          <Col span={4}>
-            <Input
-              placeholder={this.props.label}
-              value={value}
-              onChange={this.onChange}
-            />
-          </Col>
-          <Col span={4}>
-            {this.props.children(value || this.props.defaultValue)}
-          </Col>
-        </Row>
-      </React.Fragment>
-    )
-  }
+  return (
+    <React.Fragment>
+      <Row>
+        <Col span={4}>
+          <Input placeholder={props.label} value={value} onChange={onChange} />
+        </Col>
+        <Col span={4}>{props.children(value || props.defaultValue)}</Col>
+      </Row>
+    </React.Fragment>
+  )
 }
